Update output and download check
The txt files for each analysis are now in one location. If one module is run, the txt file location is the same for each one. Output for station coordinates and properties written. Curently finishing up download check for countries. It is finished for state/province downloads. This check ensures that only stations with the elements of interest are downloaded. Prior to this, the program downloaded all stations listed after the first filter section was completed. It then rejected stations that did not contain the correct datasets after they were parsed and downloaded. This speeds up the downloads.